working on vehicle and command handlers

+++ Vehicle/DL/Models/Vehicles/CQRS/Commands/VehicleDistance.cs
@@ -4,11 +4,11 @@ namespace VehicleDomain.DL.Models.Vehicles.CQRS.Commands;
 public class AddDistanceToVehicleDistance : ICommand
 {
     public int Id { get; private set; }
-    public int DistanceToAdd { get; private set; }
+    public double DistanceToAdd { get; private set; }
 }
 
 public class ResetVehicleMovedDistance : ICommand
 {
     public int Id { get; private set; }
-    public int NewDistance { get; private set; }
+    public double NewDistance { get; private set; }
 }
